splitting down methods of stat tracker into win percentage modules
